Eliminate extra files and scripts
