+++ .gitignore
@@ -2,3 +2,4 @@
 **/_build
 .vscode/
 .DS_Store
+.env
