--- books/template_book/images/myAssignment_page_12.txt
@@ -1,41 +1,0 @@
-‎https://islamopedia-books--pr13-feature-11-albidaya-o...‏ باب١‏ -- الرحیق المختموم
-‏کے ساس ام کو خا سے کے کھوڑے عپیشی ہے گے۔ فو کے کے کہ میس نے اپنے پروددگار کی یاد سے (غافل ہوکر) مال کی محبت ا قیار گی۔ یہاں تک کہ
-(آخاب) پدے مس جچ پ گیا۔ ‎Ft Fk LA al ELLA A uit okt Fo Ast Le weg lo (SLs)‏
-گی اور ان کے تخت پر ایک دھڑ ڈال دیا پھر اغہوں نے (خحداکی طرف) رج کیا (اور) ڈعاکی کہ سے ‎EB boo,‏ مخفرت کر اور ے الکی ‎Sb Gt‏
-میرے بعد یکو خایاں نہ ہو۔ بے لک نے بڑا عطا فرانے والا ہے۔ پھر جم نے ہوا کو ان کے مائع فرا نکر دیاککہ جہاں دہ ینا چاسے ان کے عم سے نرم ترم
-لئے گن ‎Susy wt‏ & (ان کے زیر ‎Boy ew (by Sue‏ بنانے وانے اور خحوطہ مارنے والے تھ۔ اور اوروں کو بھی جو ز نیروں میں ‎ane B®‏
-‎bee (WL ADE‏ شش سے ‎(xe)‏ اصا نکر دا (چاہوق) رکہ جھوڑہ (م سے) یجھہ صاب ‎Ey Be we of‏ ہمارے ہاں قرب اور
-‎a el oF‏ (۳۸:٣۳۔۰م)‏
-‏ان آیات ‎ol Ss LOU Ut a2‏ علیہ اللام جیا فرزند عطا ‎Sits ont LLU‏ کے ان ‎aut BV uk & ete A Fk FE‏
-ایک تو ىہ کہ وہ ہر عالت میں اللہ تما ی کے مائع فرمان تھے اور ہر عال جس اس ‎Sy eres BES Ewe‏ جانوروں کے ساجقھ جس طرح
-حفقت سے چیی آتے تے۔ اللہ تعالی نے گھوڑوں کی ناگوں اور گردنوں پر جو ان کے ہاتھ پچھیرنے کا ذکر فرمایا ان کے پارے میں ‎KP‏ مفسرین کے ہیں کہ وہ
-‎Ft Sot et‏ گھردنوں کے پال کاٹا کرتے تھے اور من نے کہا ہے کہ وہ ان دونوں جبوں سے ‎ZL Bee UF et od Ful‏
-اس کے علادہ اللہ تی نے ب بھی ‎i St‏ میں بھی ان کی نماز عصر قضا ہو جائی شی کیوککہ سور غروب ہو جاتا تھا۔ ىہ بات حضرت ‏ لی رشی
-ال ‎BY uv SBE A HT br Loe ot tbe WL‏ ہو جانے کا ذکر بھی کیا سے اور ہہ فرمایا سے کہ ترک نماز وی تو
-کی جات میں جائتز ‎ae SB tyne br (sth br LBM‏ سے پا بلک کے موتع پر گھوڑوں کی تیاری اور ان کا معات
-کرت ہوۓ وقت عص مگزر جاۓ تو نماز قضا کر ینا جبوراً جوا نکی ححثیت رکتا ے۔
-‎a‏ شاپ نے بھی حطرت علی رضی الثر عدہ کے اس قول کی ‎UF a Ft‏ اور اوذائیٌ نے اس بارے مس شدید چنگک کے موا قح کی شر جن ی کی ے۔ یم
-نے سورءٴ ضاء کی تفی ‎ee CEE Bin Lt‏
-‏این جرب نے ان آیات مل ارخاد پلای تعالی کہ ‎Fut uk‏ اگوں اور گرونوں پر ہاتھ پیر اکرتے سے کے بارے میں دو وضاتجی اقوال یں سے کہ
-لماع نے ان کی ٹاگوں اور گمرونوں سے پسینہ صاف کیا کرتے تے پا فگوار سے ان کی ٹاگوں کے پال صاف کیا کرتے تے اور ان کی گروثوں کے پال یا ان
-سے اجن (یر) کا ٹا کرت تھ اؤل الک قو ل کو اخقیار کیا سے اود ا کی وجہ ىہ بتائی ہے کہ دوسری وضاحت کے فحاظ سے ہہ لام آتا سے کہ ا طر چانوروں
-کو اذیت دی جاے۔ البتہ وہ بھی کے ہیں کہ گھوڑوں کو تل ف کرنا یچنی ان کو مار ڈالنا صرف اس صورت میں چائز ہو تا ےکہ جب بے خوف ہو کہ وہ زثرہ رے
-‎Sie Le GAL os‏ ىی اندینے کی وج سے ایک موقع پر جعفر بن الی طالب نے اپنے کھوڑے کو موت کے گحاٹ اتا دیا تل
-‎ek‏ کے گھوڑوں کی تعداو لت روایات میں ٹیس برار اور لحض میں دس برار بتالی گنی سے اور ہہ بھی بتایا گیا سے کہ ان میں سے ٹیں گھوڑے پر وار تھے
-‏الو دا نے اپیقی تاب سفن میں دوسرے متمعدد راویوں کے حوالے سے حرت عائشہ رضی اللہ عنہ کی زبانی ىہ عدیث بیان گی سے کہ غزوة تجوک کے سفر میں
-کی کی وجہ سے انبوں نے اپقی عوادری کا پر دہ ایک طرف سے وا سا بھول رکھا تھا کہ ہآ حضرت خفطرنے ا ن کی سوادری کے قریب اکر ال کا وچ دریافت
-‎Feud TY tu Lot “= Su dy ee ot oe Le UF) thee 3‏ نہیں ‎"od,‏
-‏صخرت عائقہ رضی 2 عزاعا بولیں: وی فو ہیں لیگن ایک نو یہ کہ میں نے اتی ‎Est OEE Fe i td‏ دوسرے پہ کہ
-‎er‏ اللہ عنہ ‎a Pet Sy peu Sud‏ وسلم ‎Wb‏ دا رگھوڑے! کیا ہیں کھوڑے بھی پردار ہوتے
-‎i te Pe Sy Jr y 627 "Su‏ بولیں: یں نے سنا سے کہ سماع کے کھوڑوں میں ھہ گھوڑے پر دا ربھی ہوتے تھے“
-‏صخرت عائشہ رشی اللہ عنہ فی ہیں کہ ان کا ہہ جواب سس نکر آحضرت ٹس پڑے جس سے آپ کے وندان مباک نظ رآنے گے۔
-‎Ae Se Ok ele‏ الد کے نام پ اپنے پا مھوڑ دتے ہیں تو اللد تعاٹی ان کے بدنے میں انی ں کوگی الصی ہز عطا فرما دیتا سے جو ان کے
-لے ان سے زیدہ یتر خابت ہوقی ہے جب کہ دہ چا جو اس کے نام پر بچوڑے جاتے ہیں وہ آزادی سے بھی کہیں پھرتے رج ہیں جاکہ ان کی تعداد زیادہ
-‎BD ta? Lad ot he ot Ket ot Sup Lai cu ts Ep PSK LI ‏پر جم ان شام اللہ‎ Bu
-‏دہ اور ای ‎Mi Us i AT Se Lie Lew‏ سخ ریا کرتے تھے ‎Lit ke owt Lui Sok‏ بدہ آیا اور ان سے کہا کہ ر ول
-‏لد میم نے ایک روز اس کا باتھ اپنے دست مبارک میں نے کر فرمایا کہ ج بکوئی شس کوئی یز خداکی راہ میں فدہ کر دیتا سے تذ خداے تعالی اسے ال کے
-بدنے میں کوئی ای چز عطا فرا دیتا سے جھ اس کی چز سے بر ہوتی ے۔ لیکن اللہ تی سے بلا ضرورت کوئی چیز مت ماگ ھکیوککہ وہ اس چز میس جھ اس نے
-‎wl Sue ‏تمہاری‎ FA bes ke?
-‏م نے سلمانع کے ضے میں آیت قرآنی خس میں بیان کیا گیا ہے کہ انہوں نے اپنے تخت پر ایک بے س رکا دعڑ پڑا بایا تھا بیہاں صرف کلام لی سے حوالے
-کے سے می کی سے جس کے بارے مس این جریر اور این ای عاتم کے علاوہ تنعدد دنر مفسرین نے بہت کچھ کھھا سے نیز اس کے متحلق اسرائیلات میں بھی
-بہت یآ کہا گیا ہے جو ظاہر ہے کہ اخویات و خرافات سے پر سے اود ای سے قابل اعتاد نہیں ہے۔ ماہم ہم نے اس موضوع پر اپنی زاب تضیر میں تتہلی کن
