+++ scripts/text_extractor/README.md
@@ -1,0 +1,5 @@
+# import os
+Imports Python’s built-in os module, which provides functions to interact with the operating system, such as working with file paths, directories, and environment variables.
+
+# import fitz
+Imports the fitz module from PyMuPDF, which is used to read, manipulate, and extract content (like text or images) from PDF files and other supported documents.
