+++ scripts/text_extractor/requirments.txt
@@ -1,0 +1,1 @@
+PyMuPDF==1.23.7
