--- uploads/chap_1.md
@@ -1,27 +1,0 @@
-## ابتدائی حالات اور نسب نامہ
-
-نبی اکرم ﷺ کا خاندان قریش میں شمار ہوتا ہے، جو حضرت ابراہیم علیہ السلام کی نسل سے تھے۔ آپ ﷺ کا سلسلہ نسب کچھ اس طرح ہے:
-
-محمد بن عبداللہ بن عبدالمطلب بن ہاشم بن عبد مناف بن قصی بن کلاب بن مرہ بن کعب بن لوی بن غالب بن فہر بن مالک بن نضر بن کنانہ بن خزیمہ بن مدرکہ بن الیاس بن مضر بن نزار بن معد بن عدنان۔
-
-حضرت ابراہیم علیہ السلام کی اولاد میں حضرت اسماعیل علیہ السلام وہ فرزند ہیں جو ان کے ساتھ مکہ میں رہے اور یہیں سے نبی اکرم ﷺ کا نسب شروع ہوتا ہے۔
-
-## قریش کے سرداروں کا ابو طالب کے پاس آنا
-
-قریش کے سرداروں کو جب یہ احساس ہوا کہ ابو طالب کی عمر پوری ہونے کے قریب ہے، تو انہیں اندیشہ ہوا کہ اگر وہ اس دنیا سے چلے گئے تو محمد صلی اللہ علیہ وسلم کا ان کی طرف سے کوئی دفاع کرنے والا نہ رہے گا۔ لہٰذا، قریش کے سردار ابو طالب کے پاس ایک بار پھر آئے اور ان سے درخواست کی کہ وہ اپنے بھتیجے سے گفتگو کریں تاکہ وہ ہمارے خداؤں کی برائی کرنا اور ہمارے دین کو چھوڑنے کی باتیں کرنا بند کریں۔
-
-وہ لوگ ابو طالب کے پاس جمع ہوئے اور کہا:
-
-"اے ابو طالب! آپ ہمارے بزرگ اور سردار ہیں۔ آپ کو معلوم ہے کہ آپ کی موت کا وقت قریب ہے۔ ہم چاہتے ہیں کہ آپ اپنے بھتیجے سے کہیں کہ وہ ہمارے خلاف جو باتیں کرتا ہے انہیں چھوڑ دے اور ہمیں ہمارے دین پر چھوڑ دے۔"
-
-اس پر ابو طالب نے رسول اللہ صلی اللہ علیہ وسلم کو بلایا اور ان کے سامنے قریش کے سرداروں کی بات رکھی۔ رسول اللہ صلی اللہ علیہ وسلم نے فرمایا:
-
-"اگر یہ لوگ میرے ایک جملے کو مان لیں تو یہ ان پر عرب اور عجم دونوں کا قبضہ ہو جائے گا۔"
-
-قریش کے سرداروں نے حیرت سے پوچھا: "وہ ایک جملہ کیا ہے؟"
-
-رسول اللہ صلی اللہ علیہ وسلم نے فرمایا:
-
-"کہو لا الٰہ الا اللہ (اللہ کے سوا کوئی معبود نہیں ہے)۔"
-
-یہ سن کر قریش کے سردار ناراض ہو گئے اور وہاں سے اٹھ کر چلے گئے۔
